requirements.txt
python-telegram-bot[rate-limiter]==22.3
pymongo==4.8.0
motor==3.5.1
python-dotenv==1.0.1
aiofiles==24.1.0
asyncio-throttle==1.0.2
sentry-sdk==2.13.0
gunicorn==23.0.0
uvloop==0.20.0
ujson==5.10.0
